+++ lib/cldr/backend/interval.ex
@@ -73,7 +73,7 @@ defmodule Cldr.Interval.Backend do
             for each of the date and time part is used
 
         * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
-          or a `Cldr.LanguageTag` struct.  The default is `#{backend}.get_locale/0`
+          or a `t:Cldr.LanguageTag.t/0` struct.  The default is `#{backend}.get_locale/0`
 
         * `number_system:` a number system into which the formatted date digits should
           be transliterated
@@ -184,7 +184,7 @@ defmodule Cldr.Interval.Backend do
             for each of the date and time part is used
 
         * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
-          or a `Cldr.LanguageTag` struct.  The default is `#{backend}.get_locale/0`
+          or a `t:Cldr.LanguageTag.t/0` struct.  The default is `#{backend}.get_locale/0`
 
         * `number_system:` a number system into which the formatted date digits should
           be transliterated.
@@ -315,7 +315,7 @@ defmodule Cldr.Interval.Backend do
             for each of the date and time part is used.
 
         * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
-          or a `Cldr.LanguageTag` struct.  The default is `#{backend}.get_locale/0`.
+          or a `t:Cldr.LanguageTag.t/0` struct.  The default is `#{backend}.get_locale/0`.
 
         * `number_system:` a number system into which the formatted date digits should
           be transliterated.
@@ -427,7 +427,7 @@ defmodule Cldr.Interval.Backend do
             for each of the date and time part is used.
 
         * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
-          or a `Cldr.LanguageTag` struct.  The default is `#{backend}.get_locale/0`.
+          or a `t:Cldr.LanguageTag.t/0` struct.  The default is `#{backend}.get_locale/0`.
 
         * `number_system:` a number system into which the formatted date digits should
           be transliterated.