+++ lib/cldr/backend/interval/date.ex
@@ -62,7 +62,7 @@ defmodule Cldr.Date.Interval.Backend do
           and `:year_and_month`. The default is `:date`.
 
         * `:locale` is any valid locale name returned by `Cldr.known_locale_names/0`
-          or a `Cldr.LanguageTag` struct.  The default is `#{backend}.get_locale/0`
+          or a `t:Cldr.LanguageTag.t/0` struct.  The default is `#{backend}.get_locale/0`
 
         * `:number_system` a number system into which the formatted date digits should
           be transliterated
@@ -176,7 +176,7 @@ defmodule Cldr.Date.Interval.Backend do
           and `:year_and_month`. The default is `:date`.
 
         * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
-          or a `Cldr.LanguageTag` struct.  The default is `#{backend}.get_locale/0`
+          or a `t:Cldr.LanguageTag.t/0` struct.  The default is `#{backend}.get_locale/0`
 
         * `number_system:` a number system into which the formatted date digits should
           be transliterated
@@ -281,7 +281,7 @@ defmodule Cldr.Date.Interval.Backend do
           and `:year_and_month`. The default is `:date`.
 
         * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
-          or a `Cldr.LanguageTag` struct.  The default is `#{backend}.get_locale/0`
+          or a `t:Cldr.LanguageTag.t/0` struct.  The default is `#{backend}.get_locale/0`
 
         * `number_system:` a number system into which the formatted date digits should
           be transliterated
@@ -392,7 +392,7 @@ defmodule Cldr.Date.Interval.Backend do
           and `:year_and_month`. The default is `:date`.
 
         * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
-          or a `Cldr.LanguageTag` struct.  The default is `#{backend}.get_locale/0`.
+          or a `t:Cldr.LanguageTag.t/0` struct.  The default is `#{backend}.get_locale/0`.
 
         * `number_system:` a number system into which the formatted date digits should
           be transliterated.