+++ lib/cldr/backend/interval/time.ex
@@ -53,7 +53,7 @@ defmodule Cldr.Time.Interval.Backend do
           and `:flex`. The default is `:time`.
 
         * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
-          or a `Cldr.LanguageTag` struct.  The default is `#{backend}.get_locale/0`.
+          or a `t:Cldr.LanguageTag.t/0` struct.  The default is `#{backend}.get_locale/0`.
 
         * `number_system:` a number system into which the formatted date digits should
           be transliterated.
@@ -163,7 +163,7 @@ defmodule Cldr.Time.Interval.Backend do
           and `:flex`. The default is `:time`.
 
         * `locale` is any valid locale name returned by `Cldr.known_locale_names/0`
-          or a `Cldr.LanguageTag` struct.  The default is `#{backend}.get_locale/0`.
+          or a `t:Cldr.LanguageTag.t/0` struct.  The default is `#{backend}.get_locale/0`.
 
         * `number_system:` a number system into which the formatted date digits should
           be transliterated.